+++ src/main/java/gregicadditions/machines/multi/advance/MetaTileEntityAdvFusionReactor.java
@@ -260,7 +260,7 @@ public class MetaTileEntityAdvFusionReactor extends RecipeMapMultiblockControlle
                     textList.add(new TextComponentTranslation("gregtech.multiblock.work_paused"));
                 } else if (this.recipeMapWorkable.isActive()) {
                     textList.add(new TextComponentTranslation("gregtech.multiblock.running"));
-                    int currentProgress;
+                    long currentProgress;
                     if (energyContainer.getEnergyCapacity() > 0) {
                         currentProgress = (int) (this.recipeMapWorkable.getProgressPercent() * 100.0D);
                         textList.add(new TextComponentTranslation("gregtech.multiblock.progress", currentProgress));
